Renamed view package to control
com.github.fishio.view --> com.github.fishio.control
